… scripts/benchmarks/refs/a100_80gb.csv